+++ src/constants/index.ts
@@ -3,3 +3,6 @@ export const SECOND = 1000;
 
 export const TOAST_TIMEOUT = 30 * SECOND;
 export const pollInterval = 60 * SECOND;
+export const PAGE_SIZES = [10, 20, 50, 100];
+export const RECENT_PAGE_SIZE_KEY = "recentPageSize";
+export const DEFAULT_PAGE_SIZE = 10;
