+++ src/hooks/useFilterInputChangeHandler.ts
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import debounce from "lodash.debounce";
 import { useLocation, useHistory } from "react-router-dom";
-import { url } from "utils";
-import { parseQueryString } from "utils/queryString/parseQueryString";
+import { url, queryString } from "utils";
 
+const { parseQueryString } = queryString;
 const { updateUrlQueryParam } = url;
 
 const updateQueryParamWithDebounce = debounce(updateUrlQueryParam, 250);
