+++ src/pages/PatchRedirect.tsx
@@ -1,7 +1,9 @@
 import { useParams, useLocation, Redirect } from "react-router-dom";
 import { getVersionRoute } from "constants/routes";
 import { PatchTab } from "types/patch";
-import { parseQueryString } from "utils/queryString";
+import { queryString } from "utils";
+
+const { parseQueryString } = queryString;
 
 export const PatchRedirect = () => {
   const { id, tab } = useParams<{ id: string; tab: PatchTab }>();
